requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1
pyyaml==6.0.2
httpx==0.28.1x`
orjson==3.10.12
SQLAlchemy==2.0.36
alembic==1.14.0
# passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
cloudinary==1.44.0
google-auth==2.35.0
firebase-admin==6.5.0
pytest==8.3.4
pytest-asyncio==0.24.0


PyJWT
passlib==1.7.4
bcrypt==3.2.2
email-validator